requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
pydantic==2.5.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
Pillow==10.1.0
